+++ src/screens/SignInEmail.tsx
@@ -68,7 +68,7 @@ const Login: React.FC<SignInEmailProps> = props => {
     <View style={classes.container}>
       <StatusBar barStyle='default' />
       <View style={classes.titleContainer}>
-        <Text style={[classes.text, classes.textTitle]}>Welcome back!</Text>
+        <Text style={[classes.text, classes.textTitle]}>Welcome!</Text>
         <Text style={[classes.text, classes.textSubtitle]}>
           We're so excited to see you again!
         </Text>
@@ -76,20 +76,20 @@ const Login: React.FC<SignInEmailProps> = props => {
       <CustomInput
         style={{
           color               : COLORS.WHITE,
-          icon                : 'person',
           marginTop           : 40,
           placeHolder         : 'Email',
           placeHolderTextColor: COLORS.LEAD,
           size                : 16
         }}
+
         value={email}
         onChangeText={handleOnChangeEmail}
         keyboardType='email-address'
+        icon='person'
       />
       <CustomInput
         style={{
           color               : COLORS.WHITE,
-          icon                : 'lock',
           placeHolder         : 'Password',
           placeHolderTextColor: COLORS.LEAD,
           size                : 16
@@ -97,6 +97,7 @@ const Login: React.FC<SignInEmailProps> = props => {
         secureTextEntry={true}
         value={password}
         onChangeText={handleOnChangePassword}
+        icon='lock'
       />
       <CustomButton
         hasIconRight={true}