+++ changelog.md
@@ -5,6 +5,11 @@
 - Implemented:
   - Validation for sign in with email view.
 
+## Version 0.15.0
+
+- Implemented:
+  - Firebase configuration was set.
+
 ## Version 0.14.0
 
 - Implemented:
@@ -17,6 +22,9 @@
   - Some changes for adding typing functionality to text fields on SignUp flow.
   - Validations for text fields on SignUp flow.
   - Reformat on screen props
+  - UI of the Chat screen of the [design](https://www.figma.com/file/qDfkoP6JJnu4f19ZmrW5Xo/IHC-Project?node-id=0%3A1).
+- TODO:
+  - Fix the bug when the keyboard is open.
 
 ## Version 0.12.0